+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Bumped the minimum required version of `joblib` to v1.3.0,
   which is the version where `joblib.parallel_config` was introduced and
   `joblib.parallel_backend` was deprecated. (#488)
+- Improved the startup time of `import civis` with a 5x speed boost. (#490)
+- The downloaded API spec due to the `civis.APIClient` instantiation is now
+  a time-to-live cache in memory (15 minutes for interactive Python, or 24 hours in scripts). (#491)
 
 ### Deprecated
 ### Removed